RevisionDiff Update
Refresh of the FileStatusList used in various forms to more consistently use Revision and ParentGuid
Allow diff to many in diff Tab
Allow multiselect operations for Submodule operations

Fixes #4485, fixes #4484, fixes #4396, fixes #4387
TBD issue to multiselect revisions and see diffCommitSubmoduleChanges
Not implemented: "Switch selected revisions"
Not implemented: Replace assigning of FileStatusList.GitItemStatus with SetDiffs() for other modules (so Revision and Parent is consistently set)

Tests are are lame, they are just adding maintenance as they are not testing the capabilities of the actions.

WIP, may split this up slightly.