Set git-status default to --ignore-submodules=none (#4760)
* Set git-status default to --ignore-submodules=none

Fixes #4754
Explicitly always include the Git command default (as it is right now). Should be a similar problem for other situations where the command is used to. 
This also fixes a problem that would occur if a user have set config option status.submoduleSummary, the output would not be the desired

* Add a test for ignoring submodules option to GetAllChangedFilesCmd()

* Explicitly use --ignore-submodules=all instead of relying that this is the same as --ignore-submodules