FileStatusList: Assume 4 selections are two ranges
When selecting 3 or 4 commits show common/unique diff if common base
(when first/last is in different branches)

If no common base, show multi diff (as is currently)

For 4 assume that selection is done in ranges, i.e. baseA..headA and baseB..headB
where A and B are in separate branches.
Therefore use the second selected commit as "first"
This is a preparation for comparing ranges

Refactoring with early return