fix some builds result not displayed
because early exit in `OnBuildInfoUpdate()`
due to `_buildStatusCancellationToken == null`.

Indeed, if the build results arrived immediatly,
`_buildStatusCancellationToken` is still not initialized
(because it will be when object is created)
when the first results will be processed.

The solution is to `lock(){}` so that it insure
that `_buildStatusCancellationToken` is set
when processing the first build result...

Cache finished builds results (to prevent a costly call at every grid revision refresh)

AzureDevOps CI: Take finished builds from cache

and reduce scope of the first api call that retrieve all the builds

Consequences:

* Immediate build results displayed
* First call is very light and quick

AzureDevOps CI: Keep in cache only the last finished build per commit

Destroy build cache when moving to another repository

By settings a sort of "watchdog" flag.

If we detect that we still are in the same repository
and so that we still need the cache, we set the flag.
That prevent the cache to be disposed when the previous
instance of a `IBuildServerAdapter` is disposed in
`BuildServerWatcher.LaunchBuildServerInfoFetchOperationAsync()`.

If we don't succeed at setting the flag on time (because
we switched of repository and so we don't use the `AzureDevOpsAdapter`),
then the cache is disposed :fireworks:

BuildServerIntegration: remove forgotten file TfsAdapter.cs

Follow up of 8190d0c0570af95ffbf9e168072d1b169eb99618

AzureDevOpsAdapter: refactoring/cleaning

make logic more explicit & add comments

AzureDevOpsAdapter: remove redundant checks

if we are here `CacheAzureDevOps` and `_buildDefinitions` are not null

BuildServerAdapter: add a new method called when changing from repository

to remove the hack done to detect the switch between repositories
to be able to clean the builds cache

BREAKING CHANGE: Change the `IBuildServerAdapter` interface by adding a new method
and so will break the external BuildIntegration plugins
(like https://github.com/martin211/GitExtensions.GitLabPipelinesPlugin ).
They will require a new release (but maybe it's the good moment to do so
because they will anyway have to do a new release built with .net5

Refactoring: Some more cleaning/renaming