Custom difftool in RevDiff, FileHistory, Commit
The default behaviour is not changed, it is still possible to click the
dropdown "parent" item to enable the functionality.

The dropdown is not shown by default for entries existing in GE to not
confuse users and to show the shortcut (that is hidden when a dropdown
is enabled).
The dropdown is toggled with a rightclick.
A tooltip is added to enable discovery.

Note that for FormFileHistory the F3 shortcut is disabled as the
shortcut is not handled as a hotkey and a dropdown disables the shortcut.

Show the default tool first and checked in the dropdown.
This is also changed for FormResolveConflicts.

Remove the separate custom difftool menu in RevDiff.