GitRevision: Remove Name property
Name was used in FormFileHistory for a specific file only,
not fitting the general use in RevisionReader.
Instead just parse this data in FormFileHistory where it is used.

* Displaying a file.
  - Set title (if it differs).
  - Check if the file exists on the filesystem (affects operations on submodules and line patching).
* Diff a file, select the old filename in an older revision.
* Use old filename in SaveAs dialog.

The parsing in StreamExtensions was not reliable for git-log -z --name-only,
requiring searches for  double null terminated records
as well as some guessing for commits without filnames (like stash).
Parse errors had to be silently ignored.

When skipping the generic case (and variable length commit message),
this is much simpler.

This allowed for removing EndOfBody handling in RevisionReader.