Support Git in WSL
For Git repos stored in WSL2 directories, execute the Git executable in WSL
where possible to improve performance. WSL Git commands are several
times faster when executed by the WSL Git executable instead of the
Git Extensions native (Windows) executable.

Using the WSL executable requires that paths provided to WSL Git
are adjusted from Git Extensions native/internal (Windows) format to WSL, like
\\wsl$\Ubunto\repo to /repo and c:\repo to /mnt/c/repo.
The paths internal to Git Extensions are always native (Windows).

The Git Extensions native (Windows) Git executable is still used for the
following:
* All handling and settings related to Git in Settings. This includes
display of Git version as well.
However, if the Git version is too old to be supported, the GitVersion
check added in #9680 will report that the version is too old. (This
change is in a separate commit.)
* Custom merge implementation in FormResolveConflicts
* Custom difftool/mergetool, the tools are only extracted in Windows.
Default git-difftool is using the Git executable, which could be separate
for WSL Git.
* Some built-in plugins etc, ScriptRunner and FindLargeFiles always use
 native Git