fixup! Revert resetting side panel at "repo changes"
RepoChangedNotifier.Notify() is used in over 30 calls situations including
the refresh button.
repo.Rest() will cause the panel to be reset in these situations which
is more annoying than not clearing the side panel immediately when switching repo.

Sidepanel is cleared if "structure changed" quickly after
the async changes anyway.

The handling should probably be changed to detect if the repo
has been changed externally...