Reorder and group items in View menu
Group the menu items to reflect how the items are used,
separated with group headings

A number of related fixes:

* stash and notes could be shown with current and filtered branches
This normally did not occur as they were filtered, but it affected tests

* Decouple ShowMergeCommits and ShowRevisionGridGraphColumn

* Reflog was not coupled with other branch revision filters
Even if Reflog dominated the other branch revision filters,
this was not reflected in the menus.
Also toggling the reflog (with the button) would set the branchfilter
to all branches.

Add reflog to branch filter dropdown, to show current filter