Custom options to git-log
If "Message" starts with "--not", interpret the string as options
to git-log.
This enables the possibility to enter filters that the GUI
does not support (like negating conditions).

The options has no checks for Git synax or if ParentsAreRewritten etc
need to be adjusted, it is up to the user.