feat: Avoid coloring moved lines
Only parse the document for added/removed lines once, by using
the DiffLineInfo info parsed for the linenumbering.
Previously, the document was parsed many times every time the
highlighting (including searching and marking) was changed.
This simplifies the code.

The filenames in the header is currently not highlighted.

Use heuristics to select which lines that are candidates to
mark "insignificant" changes for, avoid files that Git has detected
as moved files.